CMake-ified build system
The old build system is extremely fragile and wretched. This is a lot
cleaner.

Instead of concatenating files together, we put the includes in the
files themselves. Weird, right? MRC had a major distaste for ifdef, but
managed to make something worse. Since we don't care for anything other
than Unix or Windows (though I have no problem if someone wants to add
them and fit it into the new build system), we can clean house here.

Still to do:
- Windows support (osdep cleanup needed)
- Kerberos support (make dynamic in auths et al)
- Testing
- Cleaning up the rest of this codebase (and removing server-only bits)